// src/lib/social-types.ts
export type ContentFormat =
  | "Feed 1080x1350"
  | "Feed quadrado 1080x1080"
  | "Story 1080x1920"
  | "Reels 1080x1920"
  | "Carrossel 5 páginas"
  | "Carrossel 8 páginas"
  | "Facebook 1200x630"
  | "Vídeo curto"
  | "Thumbnail";